# specs/007-client-management-heartbeat/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: MAGI Device Client Management Contract
  version: 0.1.0
  description: >
    Contract for account-owned device clients, automatic TV registration,
    rotating refresh tokens, and foreground heartbeat. RFC 8628 authorization
    remains available for legacy clients; existing open read operations and
    confidential Client Credentials remain backward compatible.
servers:
  - url: /
...
paths:
  /api/open/v1/auth/device-register:
    post:
      tags: [Device Authorization]
      operationId: registerDefaultDevice
      summary: Register a TV to the configured default account
      description: >
        First-run TV clients use this endpoint without a user-entered code.
        The server resolves the configured default account, creates or reuses
        the installation record, and returns rotating device credentials.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceRegistrationRequest"
      responses:
        "200":
          description: Device credentials issued
          headers:
            Cache-Control:
              schema:
                type: string
              description: Must be no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenEnvelope"
        "400":
          $ref: "#/components/responses/ValidationProblem"
        "401":
          $ref: "#/components/responses/InvalidPublicClient"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/open/v1/auth/device-authorization:
    post:
      tags: [Device Authorization]
      operationId: beginDeviceAuthorization
      summary: Start a short-lived TV device authorization
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceAuthorizationRequest"
      responses:
        "200":
          description: Device and user codes created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceAuthorizationEnvelope"
        "400":
          $ref: "#/components/responses/ValidationProblem"
        "401":
          $ref: "#/components/responses/InvalidPublicClient"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/open/v1/auth/token:
    post:
      tags: [Device Authorization]
      operationId: issueOpenAccessToken
      summary: >
        Issue an access token from confidential client credentials, an approved
        device code, or a rotating device refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenGrantRequest"
      responses:
        "200":
          description: Token issued
          headers:
            Cache-Control:
              schema:
                type: string
              description: Must be no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenEnvelope"
        "400":
          description: Pending, slowed, denied, expired, replayed, or malformed grant
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/OAuthProblem"
        "401":
          description: Invalid confidential client or revoked device
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/OAuthProblem"
        "429":
          $ref: "#/components/responses/RateLimited"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
    cookieAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
...
  schemas:
    DeviceAuthorizationRequest:
      type: object
      additionalProperties: false
      required:
        - client_id
        - device_type
        - platform
        - platform_version
        - app_version
        - identity_summary
      properties:
        client_id:
          type: string
          minLength: 1
          maxLength: 64
          description: Registered public-device OAuth software client
        device_type:
          $ref: "#/components/schemas/DeviceType"
        platform:
          type: string
          enum: [android]
        platform_version:
          type: string
          minLength: 1
          maxLength: 64
        app_version:
          type: string
          minLength: 1
          maxLength: 64
        identity_summary:
          type: string
          minLength: 1
          maxLength: 120
          description: Non-sensitive model/manufacturer summary
        suggested_name:
          type: [string, "null"]
          minLength: 1
          maxLength: 64
    DeviceRegistrationRequest:
      type: object
      additionalProperties: false
      required:
        - client_id
        - device_type
        - platform
        - platform_version
        - app_version
        - identity_summary
        - installation_id
      properties:
        client_id:
          type: string
          minLength: 1
          maxLength: 64
        device_type:
          $ref: "#/components/schemas/DeviceType"
        platform:
          type: string
          enum: [android]
        platform_version:
          type: string
          minLength: 1
          maxLength: 64
        app_version:
          type: string
          minLength: 1
          maxLength: 64
        identity_summary:
          type: string
          minLength: 1
          maxLength: 120
        suggested_name:
          type: [string, "null"]
          minLength: 1
          maxLength: 64
        installation_id:
          type: string
          format: uuid
          description: Stable random identifier retained by one TV installation
    DeviceAuthorizationResponse:
      type: object
      additionalProperties: false
      required:
        - device_code
        - user_code
        - verification_uri
        - expires_in
        - interval
      properties:
        device_code:
          type: string
          minLength: 32
          readOnly: true
        user_code:
          type: string
          pattern: "^[A-HJ-NP-Z2-9]{4}-[A-HJ-NP-Z2-9]{4}$"
        verification_uri:
          type: string
          format: uri
        verification_uri_complete:
          type: [string, "null"]
          format: uri
          readOnly: true
        expires_in:
          type: integer
          const: 600
        interval:
          type: integer
          minimum: 5
    ClientCredentialsGrantRequest:
      type: object
      additionalProperties: false
      required: [grant_type, client_id, client_secret]
      properties:
        grant_type:
          type: string
          const: client_credentials
        client_id:
          type: string
          minLength: 1
          maxLength: 64
        client_secret:
          type: string
          minLength: 1
          maxLength: 128
    DeviceCodeGrantRequest:
      type: object
      additionalProperties: false
      required: [grant_type, client_id, device_code]
      properties:
        grant_type:
          type: string
          const: "urn:ietf:params:oauth:grant-type:device_code"
        client_id:
          type: string
          minLength: 1
          maxLength: 64
        device_code:
          type: string
          minLength: 32
    RefreshTokenGrantRequest:
      type: object
      additionalProperties: false
      required: [grant_type, client_id, refresh_token]
      properties:
        grant_type:
          type: string
          const: refresh_token
        client_id:
          type: string
          minLength: 1
          maxLength: 64
        refresh_token:
          type: string
          minLength: 32
    TokenGrantRequest:
      oneOf:
        - $ref: "#/components/schemas/ClientCredentialsGrantRequest"
        - $ref: "#/components/schemas/DeviceCodeGrantRequest"
        - $ref: "#/components/schemas/RefreshTokenGrantRequest"
      discriminator:
        propertyName: grant_type
        mapping:
          client_credentials: "#/components/schemas/ClientCredentialsGrantRequest"
          "urn:ietf:params:oauth:grant-type:device_code": "#/components/schemas/DeviceCodeGrantRequest"
          refresh_token: "#/components/schemas/RefreshTokenGrantRequest"
    TokenResponse:
      type: object
      additionalProperties: false
      required: [access_token, token_type, expires_in, scope]
      properties:
        access_token:
          type: string
          minLength: 32
          readOnly: true
        token_type:
          type: string
          const: Bearer
        expires_in:
          type: integer
          const: 3600
        scope:
          type: string
        refresh_token:
          type: string
          minLength: 32
          readOnly: true
          description: Required for device_code and refresh_token grants; absent for client_credentials
        refresh_expires_in:
          type: integer
          const: 2592000
        device_client_id:
          type: string
          format: uuid
          description: Present only for device-bound grants
...
    DeviceType:
      type: string
      enum: [android_tv]
...
    Problem:
      type: object
      additionalProperties: true
      required: [code, title, status]
      properties:
        code:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: [string, object, "null"]
        requestId:
          type: [string, "null"]
          maxLength: 120
    OAuthProblem:
      allOf:
        - $ref: "#/components/schemas/Problem"
        - type: object
          properties:
            code:
              type: string
              enum:
                - unsupported-grant-type
                - invalid-client
                - client-disabled
                - client-revoked
                - authorization_pending
                - slow_down
                - access_denied
                - expired_token
                - invalid_grant
                - device-client-revoked
    DeviceAuthorizationEnvelope:
      $ref: "#/components/schemas/SuccessEnvelopeDeviceAuthorization"
    TokenEnvelope:
      $ref: "#/components/schemas/SuccessEnvelopeToken"
...
    SuccessEnvelopeDeviceAuthorization:
      type: object
      additionalProperties: false
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: "#/components/schemas/DeviceAuthorizationResponse"
    SuccessEnvelopeToken:
      type: object
      additionalProperties: false
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: "#/components/schemas/TokenResponse"
...
  responses:
    ValidationProblem:
      description: Request validation failed
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    InvalidPublicClient:
      description: Public device client is unknown, inactive, or does not allow device grant
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    RateLimited:
      description: Request rate exceeded
      headers:
        Retry-After:
          required: true
          schema:
            type: integer
            minimum: 1
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"